添加server

+++ scrm-service/src/main/java/com/huotu/scrm/service/entity/Information/Info.java
@@ -14,7 +14,7 @@ public class Info {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int Id;
+    private long Id;
 
     /**
      * 商户号
